refactor(mcp): extract rule catalog to _catalog.py
Convert mcp.py into an mcp/ package and extract RuleCatalogEntry,
_RULE_CATALOG (31 entries), and _RULE_TO_CHECK into _catalog.py
(454 lines of pure data). __init__.py drops from 938 to 507 lines.

Closes #375
